# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Contacts, Temples & Books API",
    "description": "API for managing contacts, temples, and books for CSE341.",
    "version": "1.0.0"
  },
  "host": "cse341-node-ob82.onrender.com",
  "basePath": "/api",
  "schemes":
   [ "https" ],
...
  "securityDefinitions": {
 "cookieAuth": {
    "type": "apiKey",
      "in": "cookie",
      "name": "connect.sid",
      "description": "Session authentication via cookie (GitHub OAuth)"
  }
  },
  "paths": {
    "/contacts": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get all contacts",
        "responses": {
          "200": {
            "description": "List of contacts"
          }
        }
      },
      "post": {
  "tags": ["Contacts"],
  "summary": "Create a contact",
  "security": [{ "cookieAuth": [] }],
  "consumes": ["application/json"],
  "parameters": [
    {
      "in": "body",
      "name": "contact",
      "description": "Contact object to be created",
      "required": true,
      "schema": {
        "type": "object",
        "properties": {
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "favoriteColor": { "type": "string" },
          "birthday": { "type": "string", "format": "date" }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "favoriteColor",
          "birthday"
        ]
      }
    }
  ],
  "responses": {
    "201": {
      "description": "Contact created"
    },
    "400": {
      "description": "Bad request"
    }
    
  }
}
    },
    "/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get a contact by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact found"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "tags": ["Contacts"],
        "summary": "Update a contact by ID",
         "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "firstName": { "type": "string" },
                "lastName": { "type": "string" },
                "email": { "type": "string", "format": "email" },
                "favoriteColor": { "type": "string" },
                "birthday": { "type": "string", "format": "date" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contact updated"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete a contact by ID",
         "security": [{ "cookieAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Contact deleted"
          },
          "404": {
            "description": "Not found",
             "401": { "description": "Unauthorized" }
          }
        }
      }
    },
...
  }
}